Make asset match a helper function and move to intent_compiler
